API doesn't belong in the implementations.  Pulled out to SpoutAPI.
